fix : merge 오류 해결
